Include signature when creating instance of transaction class
Any instance of the transaction class will now include a signature, but
this signature will default to None if not specified. This is done so
that it's both possible to use the transaction class for creating new
transactions, but also for verifying old ones in the future, once that
function has been implemented.